break evictions into separate component
