+++ src/app/ranking/ranking-tool/evictors/evictors.component.html
@@ -1,0 +1,3 @@
+<p>
+  evictors works!
+</p>
