+++ src/app/ranking/ranking-tool/evictors/evictors.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-evictors',
+  templateUrl: './evictors.component.html',
+  styleUrls: ['./evictors.component.scss']
+})
+export class EvictorsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
