+++ src/app/ranking/ranking-ui/ranking-ui.component.html
@@ -1,4 +1,4 @@
-<div class="ranking-ui-container">
+<div class="ranking-ui-wrapper">
   <div class="ranking-ui-search">
     <app-predictive-search
       placeholder="{{ 'RANKINGS.SEARCH_HINT' | translate }}"
