+++ src/app/ranking/ranking.module.ts
@@ -9,6 +9,8 @@ import { RankingListComponent } from './ranking-list/ranking-list.component';
 import { RankingPanelComponent } from './ranking-panel/ranking-panel.component';
 import { RankingService } from './ranking.service';
 import { UiModule } from '../ui/ui.module';
+import { EvictorsComponent } from './ranking-tool/evictors/evictors.component';
+import { EvictionsComponent } from './ranking-tool/evictions/evictions.component';
 
 export class RankingConfig {
   cityUrl: string;
@@ -27,7 +29,9 @@ export class RankingConfig {
     RankingToolComponent,
     RankingUiComponent,
     RankingListComponent,
-    RankingPanelComponent
+    RankingPanelComponent,
+    EvictorsComponent,
+    EvictionsComponent
   ]
 })
 export class RankingModule {
